requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy==2.1.0b1
pydantic[email]==2.12.5
python-dotenv==1.2.1
cryptography==46.0.4
email-validator
